+++ example/lib/async_data.dart
@@ -1,3 +1,8 @@
+// ignore_for_file: use_late_for_private_fields_and_variables
+
+import 'dart:async';
+
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
@@ -9,7 +14,7 @@ class App extends StatelessWidget {
   App({Key? key}) : super(key: key);
 
   static const title = 'GoRouter Example: Async Data';
-  final repo = Repository();
+  static final repo = Repository();
 
   @override
   Widget build(BuildContext context) => MaterialApp.router(
@@ -22,67 +27,19 @@ class App extends StatelessWidget {
     routes: [
       GoRoute(
         path: '/',
-        pageBuilder: (context, state) => NoTransitionPage<void>(
-          key: state.pageKey,
-          child: FutureBuilder<List<Family>>(
-            future: repo.getFamilies(),
-            builder: (context, snapshot) {
-              if (snapshot.hasError) {
-                return SnapshotErrorScreen(snapshot.error!);
-              }
-
-              if (snapshot.hasData) {
-                return HomeScreen(families: snapshot.data!);
-              }
-
-              return const Center(child: CircularProgressIndicator());
-            },
-          ),
-        ),
+        builder: (context, state) => const HomeScreenWithAsync(),
         routes: [
           GoRoute(
             path: 'family/:fid',
-            pageBuilder: (context, state) => NoTransitionPage<void>(
-              key: state.pageKey,
-              child: FutureBuilder<Family>(
-                future: repo.getFamily(state.params['fid']!),
-                builder: (context, snapshot) {
-                  if (snapshot.hasError) {
-                    return SnapshotErrorScreen(snapshot.error!);
-                  }
-
-                  if (snapshot.hasData) {
-                    return FamilyScreen(family: snapshot.data!);
-                  }
-
-                  return const Center(child: CircularProgressIndicator());
-                },
-              ),
+            builder: (context, state) => FamilyScreenWithAsync(
+              fid: state.params['fid']!,
             ),
             routes: [
               GoRoute(
                 path: 'person/:pid',
-                pageBuilder: (context, state) => NoTransitionPage<void>(
-                  key: state.pageKey,
-                  child: FutureBuilder<FamilyPerson>(
-                    future: repo.getPerson(
-                      state.params['fid']!,
-                      state.params['pid']!,
-                    ),
-                    builder: (context, snapshot) {
-                      if (snapshot.hasError) {
-                        return SnapshotErrorScreen(snapshot.error!);
-                      }
-
-                      if (snapshot.hasData) {
-                        return PersonScreen(
-                            family: snapshot.data!.family,
-                            person: snapshot.data!.person);
-                      }
-
-                      return const Center(child: CircularProgressIndicator());
-                    },
-                  ),
+                builder: (context, state) => PersonScreenWithAsync(
+                  fid: state.params['fid']!,
+                  pid: state.params['pid']!,
                 ),
               ),
             ],
@@ -93,78 +50,209 @@ class App extends StatelessWidget {
   );
 }
 
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({required this.families, Key? key}) : super(key: key);
-  final List<Family> families;
+class HomeScreenWithAsync extends StatefulWidget {
+  const HomeScreenWithAsync({Key? key}) : super(key: key);
+
+  @override
+  State<HomeScreenWithAsync> createState() => _HomeScreenWithAsyncState();
+}
+
+class _HomeScreenWithAsyncState extends State<HomeScreenWithAsync> {
+  Future<List<Family>>? _future;
+  List<Family>? _families;
+
+  @override
+  void initState() {
+    super.initState();
+    fetch();
+  }
+
+  @override
+  void didUpdateWidget(covariant HomeScreenWithAsync oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    // refresh cached data
+    //fetch(); // no need
+  }
+
+  void fetch() {
+    _families = null;
+    _future = App.repo.getFamilies();
+    _future!.then(
+      (families) => setState(() => _families = families), // update AppBar
+    );
+  }
 
   @override
   Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: const Text(App.title)),
-        body: ListView(
-          children: [
-            for (final f in families)
-              ListTile(
-                title: Text(f.name),
-                onTap: () => context.go('/family/${f.id}'),
-              )
-          ],
+        appBar: AppBar(
+          title: Text(
+            '${App.title}: ${_families != null ? '${_families!.length} '
+                'families' : 'loading...'}',
+          ),
+        ),
+        body: FutureBuilder<List<Family>>(
+          future: _future,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return const Center(child: CircularProgressIndicator());
+            }
+
+            if (snapshot.hasError) return SnapshotError(snapshot.error!);
+
+            assert(snapshot.hasData);
+            return ListView(
+              children: [
+                for (final f in _families!)
+                  ListTile(
+                    title: Text(f.name),
+                    onTap: () => context.go('/family/${f.id}'),
+                  )
+              ],
+            );
+          },
         ),
       );
 }
 
-class FamilyScreen extends StatelessWidget {
-  const FamilyScreen({required this.family, Key? key}) : super(key: key);
-  final Family family;
+class FamilyScreenWithAsync extends StatefulWidget {
+  const FamilyScreenWithAsync({required this.fid, Key? key}) : super(key: key);
+  final String fid;
+
+  @override
+  State<FamilyScreenWithAsync> createState() => _FamilyScreenWithAsyncState();
+}
+
+class _FamilyScreenWithAsyncState extends State<FamilyScreenWithAsync> {
+  Future<Family>? _future;
+  Family? _family;
+
+  @override
+  void initState() {
+    super.initState();
+    fetch();
+  }
+
+  @override
+  void didUpdateWidget(covariant FamilyScreenWithAsync oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    // refresh cached data
+    if (oldWidget.fid != widget.fid) fetch();
+  }
+
+  void fetch() {
+    _family = null;
+    _future = App.repo.getFamily(widget.fid);
+    _future!.then(
+      (family) => setState(() => _family = family), // update AppBar
+    );
+  }
 
   @override
   Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: Text(family.name)),
-        body: ListView(
-          children: [
-            for (final p in family.people)
-              ListTile(
-                title: Text(p.name),
-                onTap: () => context.go('/family/${family.id}/person/${p.id}'),
-              ),
-          ],
+        appBar: AppBar(title: Text(_family?.name ?? 'loading...')),
+        body: FutureBuilder<Family>(
+          future: _future,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return const Center(child: CircularProgressIndicator());
+            }
+
+            if (snapshot.hasError) return SnapshotError(snapshot.error!);
+
+            assert(snapshot.hasData);
+            return ListView(
+              children: [
+                for (final p in _family!.people)
+                  ListTile(
+                    title: Text(p.name),
+                    onTap: () =>
+                        context.go('/family/${_family!.id}/person/${p.id}'),
+                  ),
+              ],
+            );
+          },
         ),
       );
 }
 
-class PersonScreen extends StatelessWidget {
-  const PersonScreen({required this.family, required this.person, Key? key})
+class PersonScreenWithAsync extends StatefulWidget {
+  const PersonScreenWithAsync({required this.fid, required this.pid, Key? key})
       : super(key: key);
 
-  final Family family;
-  final Person person;
+  final String fid;
+  final String pid;
+
+  @override
+  State<PersonScreenWithAsync> createState() => _PersonScreenWithAsyncState();
+}
+
+class _PersonScreenWithAsyncState extends State<PersonScreenWithAsync> {
+  Future<FamilyPerson>? _future;
+  FamilyPerson? _famper;
+
+  @override
+  void initState() {
+    super.initState();
+    fetch();
+  }
+
+  @override
+  void didUpdateWidget(covariant PersonScreenWithAsync oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    // refresh cached data
+    if (oldWidget.fid != widget.fid || oldWidget.pid != widget.pid) fetch();
+  }
+
+  void fetch() {
+    _famper = null;
+    _future = App.repo.getPerson(widget.fid, widget.pid);
+    _future!.then(
+      (famper) => setState(() => _famper = famper), // update AppBar
+    );
+  }
 
   @override
   Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: Text(person.name)),
-        body: Text('${person.name} ${family.name} is ${person.age} years old'),
+        appBar: AppBar(title: Text(_famper?.person.name ?? 'loading...')),
+        body: FutureBuilder<FamilyPerson>(
+          future: _future,
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return const Center(child: CircularProgressIndicator());
+            }
+
+            if (snapshot.hasError) return SnapshotError(snapshot.error!);
+
+            assert(snapshot.hasData);
+            return Text(
+              '${_famper!.person.name} ${_famper!.family.name} is '
+              '${_famper!.person.age} years old',
+            );
+          },
+        ),
       );
 }
 
-class SnapshotErrorScreen extends StatelessWidget {
-  SnapshotErrorScreen(Object error, {Key? key})
+class SnapshotError extends StatelessWidget {
+  SnapshotError(Object error, {Key? key})
       : error = error is Exception ? error : Exception(error),
         super(key: key);
   final Exception error;
 
   @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: const Text('Page Not Found')),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              SelectableText(error.toString()),
-              TextButton(
-                onPressed: () => context.go('/'),
-                child: const Text('Home'),
-              ),
-            ],
-          ),
+  Widget build(BuildContext context) => Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            SelectableText(error.toString()),
+            TextButton(
+              onPressed: () => context.go('/'),
+              child: const Text('Home'),
+            ),
+          ],
         ),
       );
 }