+++ lib/src/custom_transition_page.dart
@@ -157,12 +157,16 @@ class NoTransitionPage<T> extends CustomTransitionPage<T> {
   /// Constructor for a page with no transition functionality.
   const NoTransitionPage({
     required Widget child,
+    String? name,
+    Object? arguments,
     String? restorationId,
     LocalKey? key,
   }) : super(
           transitionsBuilder: _transitionsBuilder,
           transitionDuration: const Duration(microseconds: 1), // hack for #205
           key: key,
+          name: name,
+          arguments: arguments,
           restorationId: restorationId,
           child: child,
         );