+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   simpler way to build error pages via the `GoRouter.errorBuilder` method
 - provide default implementation of `GoRouter.errorBuilder` to provide an error
   page without the need to implement a custom error page builder
+- TODO: added a new `NoTransitionPage`
+- TODO: added a [new `custom_error_screen` sample]()
 - TODO: new [Migrating to 2.5 section](https://gorouter.dev/migrating-to-20) in
   the docs to show how to take advantage of the new `builder` and default error
   page builder